moving padding utils to separate file to remove pytorch dependency
